% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmer_fit.R
\name{glmer_fit}
\alias{glmer_fit}
\title{A wrapper function of the glmer function.}
\usage{
glmer_fit(
  y,
  x = NULL,
  start = NULL,
  weights = NULL,
  offset = NULL,
  ...,
  estfun = FALSE,
  object = FALSE
)
}
\arguments{
\item{y}{A matrix containing response values of
each participant for each item.}
}
\description{
A wrapper function of the glmer function.
}
