% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{poly_idx_cpp}
\alias{poly_idx_cpp}
\title{Generates a matrix containing every possible score combination
where each row sum is equal to the provided esf order.}
\usage{
poly_idx_cpp(p, M)
}
\arguments{
\item{p}{The esf order of the elementary symmetric function.}

\item{M}{The number of items}
}
\value{
Every possible score combination as a matrix
}
\description{
Generates a matrix containing every possible score combination
where each row sum is equal to the provided esf order.
}
\examples{
poly_idx_cpp(5, 7)
}
