% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcm_fit.R
\name{rpcm_fit}
\alias{rpcm_fit}
\title{A wrapper function of the rpcm function.}
\usage{
rpcm_fit(
  y,
  x = NULL,
  start = NULL,
  weights = NULL,
  offset = NULL,
  ...,
  estfun = FALSE,
  object = FALSE
)
}
\arguments{
\item{y}{A matrix containing response values of
each participant for each item.}
}
\description{
A wrapper function of the rpcm function.
}
