% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcm_tree.R, R/rpcmtree.R
\docType{package}
\name{rpcmtree}
\alias{rpcmtree}
\title{Creates a rasch poisson count model tree.}
\usage{
rpcmtree(
  formula,
  data,
  reltol = 1e-10,
  maxit = 100L,
  fitting_func = rpcm_fit,
  ...
)
}
\arguments{
\item{formula}{A symbolic description of the model to be fit.}

\item{data}{A data frame containing the variables in the model.}

\item{reltol}{Argument for optim.}

\item{maxit}{Argument for optim.}

\item{fitting_func}{A fitting function for the model.
By default the rpcm_fit function is used.}

\item{...}{Arguments passed to the underlying functions.}
}
\value{
A rasch poisson count model tree based on the input data.
}
\description{
The mypackage package provides three categories of important functions:
foo, bar and baz.
}
\section{Mypackage functions}{

The mypackage functions ...
}

